Simplex noise added
I decided for the sake of efficiency that simplex noise would be better. I appropriated and ported code courtesy of Stefan Gustavson. Credit and links to his paper and code can be found in SimpleNoise.cs and the original code is in Java. My only contribution was porting it and adding octave functionality.