Trigonometry noise added
I created a noise function based on trig waves. Each axis has a series of octaves, each of which contain values for a Sin function with randomly generated constant. At the moment results are averaged between axes and put through a fade function. May be good for surfaces similar to sand dunes.